Identacion e imports

+++ Practica/Dispositivos/Computadoras/Laptop.java
@@ -1,13 +1,18 @@
 package Dispositivos.Computadoras;
 
+import java.util.Scanner;
+
 public class Laptop extends Computadora{
-  protected int peso;
+    
+    protected int peso;
     protected int duracionBateria;
-    public Laptop(String modelo, String marca, float precio, int peso, int duracionBateria){
+    
+    public Laptop(String modelo, String marca, double precio, int peso, int duracionBateria){
         super(modelo,marca,precio);
         this.peso=peso;
         this.duracionBateria=duracionBateria;
     }
+    
     public static Laptop nuevo(Scanner sc){
         String modelo;
         String marca;
@@ -25,18 +30,23 @@ public class Laptop extends Computadora{
         duracionBateria = sc.nextInt();
         return new Laptop(modelo,marca,precio,peso,duracionBateria);
     }
+    
     public void setPeso(int peso){
         this.peso=peso;
     }
+    
     public int getPeso(){
         return peso;
     }
+    
     public void setduracionBateria(int duracionBateria){
         this.duracionBateria=duracionBateria;
     }
+    
     public int getduracionBateria(){
         return duracionBateria;
     }
+    
     public String getClassName(){
         return "LapTop";
     }
